V2, Nishi #717 adding IPProgress and CPprogress alerts

+++ whp-common/src/main/java/org/motechproject/whp/common/domain/alerts/AlertThresholds.java
@@ -16,7 +16,7 @@ public class AlertThresholds {
         }
     }
 
-    public AlertThreshold getThreshold(int value) {
+    public AlertThreshold getThreshold(double value) {
         AlertThreshold threshold = null;
         for(AlertThreshold alertThreshold : alertThresholds){
             if(value < alertThreshold.getThreshold()){
