+++ whp-patient/src/main/java/org/motechproject/whp/patient/alerts/service/AllAlertProcessors.java
@@ -33,7 +33,7 @@ public class AllAlertProcessors {
     }
 
     private void processAlert(AlertProcessor alertProcessor, Patient patient) {
-        int alertValue = alertProcessor.process(patient);
+        double alertValue = alertProcessor.process(patient);
         int severity = alertConfigurations.getAlertSeverityFor(alertProcessor.alertType(), alertValue);
         patient.updatePatientAlert(alertProcessor.alertType(), alertValue, severity);
     }
